Fixes #22 - Exception on draft_destroy when has_one association is nil
`self.send(association.name)` will return `nil` if the association is empty (there are no dependents).  Currently, if the association is `has_one` and the result is `nil`, it is still wrapped in an array.  When we try to access this `nil` element in the following `.each` block, an exception is thrown.

The changes to model.rb resolve this issue.  Singular dependents are only wrapped in an array if they exist (and `nil` isn't wrapped).

The `.each` block is also only called if the dependents exist.  So, whether a `has_one` or `has_many` returns `nil`, the block will be bypassed and we avoid any exceptions.

I added an `OnlyChild` class and association to the parents to easily test this fix.  Without the fix, many of the current tests will fail.  With the fix in place, the tests will run properly again.