Removing Annoying Draftsman Things
A lot about this gem annoys me, but with this commit pushed I will be as calm as a cucumber, as trainquil as still lake
Removed code that checked if the draft has returned to it's original state, if it had, delete it.
This just confused our system and would cause the content to enter an un-approvable state. Also this logic seems handwavy, why magically delete the draft?
It makes sense in a way, but for our use case it does not make sense.
Removed logic around create drafts.
Again this makes very little sense to me, if the draft has an event of 'create' we update the parent object. Why?
Why not just keep updating the draft?
I guess if it's never been published then you want both to stay in sync? It doesn't make much sense to me.

Signed-off-by: Evan <evan@spectrum.md>